+++ inst/examples/ex-chemistry.R
@@ -1,0 +1,10 @@
+## Data from Day et al. 2011
+data("kommos", package = "folio") # Coerce to compositional data
+kommos <- remove_NA(kommos, margin = 1) # Remove cases with missing values
+coda <- as_composition(kommos, groups = 1) # Use ceramic types for grouping
+
+is_element_major(coda)
+is_element_minor(coda)
+is_element_trace(coda)
+
+is_oxide(coda)